requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.7.3
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.4.0
gunicorn==22.0.0
